[Sprint 7] TN: INT: Monitoreo: Mant: Implementación de CRUDs Elementos monitoreo

+++ src/main/java/ec/telconet/microservicios/dependencias/esquema/infraestructura/dto/ElementoPorCantonParamsReqDTO.java
@@ -15,7 +15,7 @@ import lombok.EqualsAndHashCode;
 public class ElementoPorCantonParamsReqDTO extends ElementoReqDTO {
 	private Long cantonId;
 	private String nombreCanton;
-	private Integer tipoId;
+	private Long tipoId;
 	private String nombreTipo;
 	private String detalle;
 	private String detalleValor;
